add addition condition to check result

+++ tests/cypress/plugins/index.js
@@ -43,7 +43,7 @@ module.exports = (on, config) => {
     });
 
     on('after:spec', (spec, results) => {
-        if (results.stats.failures === 0 && results.video) {
+        if (results && results.stats.failures === 0 && results.video) {
             fs.unlinkSync(results.video);
         }
     });
